+++ src/core/instruction/RTypeInstruction.cpp
@@ -57,7 +57,7 @@ void ADD::execute(RiscvCpu& cpu, InstructionOutput& instructionOutput)
     uint32_t rs1Value = cpu.getRegister(rs1);
     uint32_t rs2Value = cpu.getRegister(rs2);
 
-    int32_t result = rs2Value + rs1Value;
+    uint32_t result = rs2Value + rs1Value;
 
     cpu.setRegister(rd, result);
     cpu.setPc(cpu.getPc() + 4);
@@ -77,7 +77,7 @@ void SUB::execute(RiscvCpu& cpu, InstructionOutput& instructionOutput)
     uint32_t rs1Value = cpu.getRegister(rs1);
     uint32_t rs2Value = cpu.getRegister(rs2);
 
-    int32_t result = rs1Value - rs2Value;
+    uint32_t result = rs1Value - rs2Value;
 
     cpu.setRegister(rd, result);
     cpu.setPc(cpu.getPc() + 4);
@@ -100,7 +100,7 @@ void SLL::execute(RiscvCpu& cpu, InstructionOutput& instructionOutput)
 
     uint8_t shiftValue = getBits(rs2Value, 0, 5);
 
-    int32_t result = rs1Value << shiftValue;
+    uint32_t result = rs1Value << shiftValue;
 
     cpu.setRegister(rd, result);
     cpu.setPc(cpu.getPc() + 4);
@@ -117,8 +117,8 @@ void SLL::execute(RiscvCpu& cpu, InstructionOutput& instructionOutput)
 
 void SLT::execute(RiscvCpu& cpu, InstructionOutput& instructionOutput)
 {
-    uint32_t rs1Value = cpu.getRegister(rs1);
-    uint32_t rs2Value = cpu.getRegister(rs2);
+    int32_t rs1Value = static_cast<int32_t>(cpu.getRegister(rs1));
+    int32_t rs2Value = static_cast<int32_t>(cpu.getRegister(rs2));
 
     uint8_t result = (rs1Value < rs2Value);
 
@@ -141,7 +141,7 @@ void SLTU::execute(RiscvCpu& cpu, InstructionOutput& instructionOutput)
     uint32_t rs1Value = cpu.getRegister(rs1);
     uint32_t rs2Value = cpu.getRegister(rs2);
 
-    uint8_t result = (static_cast<uint32_t>(rs1Value) < static_cast<uint32_t>(rs2Value));
+    uint8_t result = (rs1Value < rs2Value);
 
     cpu.setRegister(rd, result);
     cpu.setPc(cpu.getPc() + 4);
@@ -162,7 +162,7 @@ void XOR::execute(RiscvCpu& cpu, InstructionOutput& instructionOutput)
     uint32_t rs1Value = cpu.getRegister(rs1);
     uint32_t rs2Value = cpu.getRegister(rs2);
 
-    int32_t result = (rs1Value ^ rs2Value);
+    uint32_t result = (rs1Value ^ rs2Value);
 
     cpu.setRegister(rd, result);
     cpu.setPc(cpu.getPc() + 4);
@@ -184,7 +184,7 @@ void SRL::execute(RiscvCpu& cpu, InstructionOutput& instructionOutput)
 
     uint8_t shiftValue = getBits(rs2Value, 0, 5);
 
-    int32_t result = static_cast<uint32_t>(rs1Value) >> shiftValue;
+    uint32_t result = static_cast<uint32_t>(rs1Value) >> shiftValue;
 
     cpu.setRegister(rd, result);
     cpu.setPc(cpu.getPc() + 4);
@@ -206,7 +206,7 @@ void SRA::execute(RiscvCpu& cpu, InstructionOutput& instructionOutput)
 
     uint8_t shiftValue = getBits(rs2Value, 0, 5);
 
-    int32_t result = rs1Value >> shiftValue;
+    uint32_t result = static_cast<uint32_t>(static_cast<int32_t>(rs1Value) >> shiftValue);
 
     cpu.setRegister(rd, result);
     cpu.setPc(cpu.getPc() + 4);
@@ -226,7 +226,7 @@ void OR::execute(RiscvCpu& cpu, InstructionOutput& instructionOutput)
     uint32_t rs1Value = cpu.getRegister(rs1);
     uint32_t rs2Value = cpu.getRegister(rs2);
 
-    int32_t result = (rs1Value | rs2Value);
+    uint32_t result = (rs1Value | rs2Value);
 
     cpu.setRegister(rd, result);
     cpu.setPc(cpu.getPc() + 4);
@@ -246,7 +246,7 @@ void AND::execute(RiscvCpu& cpu, InstructionOutput& instructionOutput)
     uint32_t rs1Value = cpu.getRegister(rs1);
     uint32_t rs2Value = cpu.getRegister(rs2);
 
-    int32_t result = (rs1Value & rs2Value);
+    uint32_t result = (rs1Value & rs2Value);
 
     cpu.setRegister(rd, result);
     cpu.setPc(cpu.getPc() + 4);
